Make header files optional.
Missing header files are now just a warning, not an error. Undeclared
functions and variables are accepted without warnings. Types still need
to be declared, though—although some more types are built in, and names
can be declared to be types in the config file with the new "typedef"
setting.