Improve printing of enums.
Print the list of constants when the enum is defined and typedef'ed at
the same time.

Don't print the list of constants more than once for a given enum (if
it's defined in one place and used in a typedef elsewhere).

Don't bother defining a Go type for the enum, since it's translated as
int everywhere it is used.