Improve comment assignment.
Using the GCC preprocessor messed up how comments were assigned to
syntax elements. For one thing, we were getting duplicate comments from
header files that were included in more than one place. So now we use a
map to deduplicate the comments.

Also, there were often big gobs of comments that had been associated
with declarations in the C file, but those declarations went away in Go.
So the comments didn't make any sense with nothing to comment on. Now we
only keep the comments that are reasonably close to something they might
be commenting on.