Cache canonical copy of top-level declarations.
When a header file is included in multiple places, the parser produces a
copy of the declarations for each time it is included. We were already
discarding the extra copies when merging the declarations from various
files, but uses of identifiers in different scopes were still referring
to different copies of the declaration, some of which didn't get
properly typechecked because they were removed from the program's list
of declarations. This commit should fix that problem.